+++ frontend/src/app/components/home/product/display/product-view/product-view.component.scss
@@ -58,10 +58,9 @@
 .search-wrapper{
   height: 50em;
   width: 50em;
+  background-clip: content-box;
   padding: 50px;
   padding-top: 40px;
-  border: 3px solid #343442;
-  border-radius: 3px;
 }
 
 @each $theme-name in $theme-names {
@@ -74,6 +73,7 @@
   display: flex;
   justify-content: center;
   margin-top: auto;
+  overflow: auto;
 }
