--- frontend/src/app/components/home/product/display/product-view/select-categories/select-categories.component.html
@@ -1,1 +1,0 @@
-<button class="select" [ngClass]="'button-' + theme" *ngFor="let cat of cats;  let i = index;" (click)="removeCategoryByClicking(i)">{{cats[i].category}}</button>
