+++ frontend/src/app/components/home/product/search-product/search-product-component-theme.scss
@@ -12,5 +12,8 @@
     h3 {
       color: $text-color;
     }
+    b {
+      color: $text-color;
+    }
   }
 }
