+++ frontend/src/app/components/home/product/search-product/search-product.component.html
@@ -1,52 +1,39 @@
 <div>
   <form (ngSubmit)="onSubmit()">
     <div class="input-grid" >
-      <app-text-input placeholder="Title" name="titleAndDescription" [(ngModel)]="criteria.titleAndDescription">
+      <app-text-input placeholder="Search Term (for Title and Description)" name="titleAndDescription" [(ngModel)]="criteria.titleAndDescription">
       </app-text-input>
     </div>
-    <fieldset class="input-grid" id="radioGrid" [ngClass]="theme">
-      <h2>Deliverable: </h2>
-      <div class="form-check">
-        <label class="form-check-label">
-          <input type="radio" class="form-check-input" name="optionsRadios" [value]="true" [(ngModel)]="criteria.deliverable">
-          <h3>Deliverable</h3>
-        </label>
-      </div>
-      <div class="form-check">
-        <label class="form-check-label">
-          <input type="radio" class="form-check-input" name="optionsRadios" [value]="false" [(ngModel)]="criteria.deliverable">
-          <h3>Not Deliverable</h3>
-        </label>
-      </div>
-    </fieldset>
 
     <div class="input-grid">
       <app-select name="category" selectName="Category" [(ngModel)]="criteria.category" (ngModelChange)="SubCategoryShow(criteria.category)"
-          [options]="['Houses', 'Pets', 'Nice Stuff']"></app-select>
+          [options]="catArray"></app-select>
       <app-select *ngIf="show" name="subcategory" selectName="Subcategory" [(ngModel)]="criteria.subcategory"
           [options]="['Houses_sub', 'Pets_sub', 'Nice Stuff_sub']"></app-select>
+          <div #parent> </div>
     </div>
 
     <div class="input-grid">
       <app-select name="type" selectName="Product Type" [(ngModel)]="criteria.type" (ngModelChange)="OfferTypeShow(criteria.type)"
           [options]="['Item', 'Service']"></app-select>
       <app-select *ngIf="isShown" name="offerType" selectName="Offer Type" [(ngModel)]="criteria.offerType"
-          [options]="['Sell', 'Rent']"></app-select>
-      </div>
+          [options]="optionArray"></app-select>
+    </div>
 
     <div class="input-grid" >
         <app-text-input placeholder="Location(City)" name="location" [(ngModel)]="criteria.location">
         </app-text-input>
+        <div class="input-grid" [ngClass]="theme" id ="deliverableGrid"><h2>Not Deliverable</h2><mat-slide-toggle [(ngModel)]="criteria.deliverable" name="Deliverable" selectName="Product Deliverable"
+          class="example-margin" [color]="color"> <h2 id="b2">Deliverable</h2></mat-slide-toggle>
+        </div>
     </div>
 
-    <div class="input-grid" [ngClass]="theme">
+    <div class="input-grid" id="priceGrid" [ngClass]="theme">
       <h2>Price Range: </h2>
       <div class='wrap' role='group' aria-labelledby='multi-lbl'>
         <input id='a' type='range' min='0' value='3000'max='3000' name="priceMax" [(ngModel)]="criteria.priceMax"/>
         <input id='b' type='range' min='0' value='0'   max='3000' name="priceMin" [(ngModel)]="criteria.priceMin"/>
       </div>
-    </div>
-    <div class="input-grid">
       <app-text-input placeholder="Minimum Price $" name="priceMin" [(ngModel)]="criteria.priceMin">
       </app-text-input>
       <app-text-input placeholder="Maximum Price $" name="priceMax" [(ngModel)]="criteria.priceMax">
