+++ frontend/src/app/components/home/product/search-product/select-categories/select-categories.component.html
@@ -1,0 +1,3 @@
+<!-- <button class="select" [ngClass]="'button-' + theme" *ngFor="let cat of cats;  let i = index;" (click)="removeCategoryByClicking(i)">{{cats[i].category}}</button> -->
+<!-- <app-select name="category" selectName="Category" [(ngModel)]="criteria.category" (ngModelChange)="SubCategoryShow(criteria.category)"
+          [options]="['Houses', 'Pets', 'Nice Stuff']"></app-select> -->
