+++ frontend/src/app/components/home/product/search-product/select-categories/select-categories.component.scss
